💄 UI: Update Sidebar Size And Animation [feature/responsive]
Changes Made:
	- Updated sidebar width and margin.
	- Added visibility toggle and shadow effect to sidebar.
	- Updated transition duration for sidebar animation.

The sidebar never had an inaccurate height according to the page because
the page is bigger than the sidebar and screen size, causing underflow.
This made the sidebar look awkward on different devices and after
scrolling.

The sidebar was changed to fit-content height to only be the required to
show all the links without extra space, this made it look better, also
added box-shadow and border-radius for better UI. I added a visibility
toggle to the sidebar so that when it is hidden, it is not focus-able or
interacted with, improving user experience. The units were changed from
pixels to rem for better responsiveness across devices.

Files Changed:
	modified:   js/sidebar.js
	modified:   styles/main.css

Date: 08/11/2025, 12:46 P.M. IST

All Done.

Signed-off-by: Kunal Kumar <kunalkumarchourasiya2021@gmail.com>