✨ Feat: Implement PHP With Auth System [feature/php-functionality]
Changes Made:
	- Changed HTML files to PHP files to add server-side functionality.
	- Added user authentication system with login and logout features.
	- Updated CSS files to enhance the styling of the new login page.
	- Added .gitignore to exclude unnecessary files from the repository.
	- Added navigation links for login and logout in the menu.
	- Improved page styling for better user experience.

I had to add PHP functionality to the website to enable user authentication,
because it was a requirement for the project. This involved converting
existing HTML files to PHP and creating new PHP files for login and
logout.

I also changed some CSS to improve the appearance of the previous pages.
I added better spacing in the menu card items, and changed the layout of
the about page for better readability and appearance. Also added a new
.gitignore file to handle a credentials file that will be used for
environmental variables in future.

Files Changed:
	new file:   .gitignore
	deleted:    about.html
	new file:   about.php
	deleted:    contact.html
	new file:   contact.php
	deleted:    index.html
	new file:   index.php
	deleted:    locations.html
	new file:   locations.php
	new file:   login.php
	new file:   logout.php
	deleted:    menu.html
	new file:   menu.php
	modified:   styles/about.css
	new file:   styles/login.css
	modified:   styles/main.css
	modified:   styles/menu.css

Date: 09/11/2025, 06:47 P.M. IST

All Done.

Signed-off-by: Kunal Kumar <kunalkumarchourasiya2021@gmail.com>